Refactor subscription handling endpoints and remove app based subscriptions
