+++ portals/developer-portal/src/controllers/apiContentController.js
@@ -27,7 +27,7 @@ const util = require('../utils/util');
 const constants = require('../utils/constants');
 const adminDao = require('../dao/admin');
 const apiDao = require('../dao/apiMetadata');
-const platformSubDao = require('../dao/platformSubscription');
+const subDao = require('../dao/subscription');
 const apiMetadataService = require('../services/apiMetadataService');
 const { shouldShowPlatformApiKeysNav, findSubscriptionTokenHeader } = require('../services/platformApiKeysNavService');
 const adminService = require('../services/adminService');
@@ -121,20 +121,20 @@ const loadAPIs = async (req, res) => {
                 metaData.applications = perApiAppList;
             }
 
-            // Load platform subscriptions for platform APIs with subscription plans (single call for all)
+            // Load subscriptions for platform APIs with subscription plans (single call for all)
             if (req.user) {
                 try {
-                    const localSubs = await platformSubDao.listPlatformSubscriptions(orgID);
+                    const localSubs = await subDao.listSubscriptions(orgID);
                     const subscribedApiIds = new Set(localSubs.map(sub => sub.API_ID));
                     for (const metaData of metaDataList) {
                         const isPlatform = metaData.apiInfo?.gatewayType === 'wso2/api-platform';
                         const hasPlans = (metaData.subscriptionPolicies || []).length > 0;
                         if (isPlatform && hasPlans) {
-                            metaData.hasPlatformSubscription = subscribedApiIds.has(metaData.apiID);
+                            metaData.hasSubscription = subscribedApiIds.has(metaData.apiID);
                         }
                     }
                 } catch (err) {
-                    logger.warn('Failed to load platform subscriptions for API listing', {
+                    logger.warn('Failed to load subscriptions for API listing', {
                         error: err.message
                     });
                 }
@@ -437,14 +437,14 @@ const loadAPIContent = async (req, res) => {
                 }
             }
 
-            // Load platform gateway subscriptions for platform APIs with subscription plans
-            let platformSubscriptions = [];
-            const isPlatformGateway = metaData.apiInfo?.gatewayType === 'wso2/api-platform';
-            const hasPlatformPlans = (subscriptionPlans || []).length > 0;
-            if (req.user && isPlatformGateway && hasPlatformPlans) {
+            // Load subscriptions for platform APIs with subscription plans
+            let subscriptions = [];
+            const isGatewayAPI = metaData.apiInfo?.gatewayType === 'wso2/api-platform';
+            const hasPlans = (subscriptionPlans || []).length > 0;
+            if (req.user && isGatewayAPI && hasPlans) {
                 try {
-                    const localSubs = await platformSubDao.listPlatformSubscriptions(orgID, { apiId: apiID });
-                    platformSubscriptions = (localSubs || []).map(sub => ({
+                    const localSubs = await subDao.listSubscriptions(orgID, { apiId: apiID });
+                    subscriptions = (localSubs || []).map(sub => ({
                         subscriptionId: sub.SUB_ID,
                         subscriptionPlanName: sub.DP_SUBSCRIPTION_POLICY?.DISPLAY_NAME || sub.DP_SUBSCRIPTION_POLICY?.POLICY_NAME || '',
                         status: sub.STATUS,
@@ -452,7 +452,7 @@ const loadAPIContent = async (req, res) => {
                         customerName: null
                     }));
                 } catch (err) {
-                    logger.warn('Failed to load platform subscriptions', {
+                    logger.warn('Failed to load subscriptions', {
                         error: err.message, orgID, apiID
                     });
                 }
@@ -492,7 +492,7 @@ const loadAPIContent = async (req, res) => {
                 providerUrl: providerUrl,
                 apiMetadata: metaData,
                 subscriptionPlans: subscriptionPlans,
-                platformSubscriptions: platformSubscriptions,
+                subscriptions: subscriptions,
                 baseUrl: '/' + orgName + constants.ROUTE.VIEWS_PATH + viewName,
                 schemaUrl: `${req.protocol}://${req.get('host')}${constants.ROUTE.DEVPORTAL_ASSETS_BASE_PATH}${orgID}/${constants.ROUTE.API_FILE_PATH}${apiID}${constants.API_TEMPLATE_FILE_NAME}${schemaFileName}`,
                 loadDefault: loadDefault,
