+++ portals/developer-portal/src/defaultContent/layout/main.hbs
@@ -20,7 +20,6 @@
     <script src='/technical-scripts/filter.js' defer></script>
     <script src='/technical-scripts/common.js' defer></script>
     <script src='/technical-scripts/delete-confirmation-modal.js' defer></script>
-    <script src='/technical-scripts/subscription.js' defer></script>
     {{#if portalConfigs.reoClientID}}
     <!-- Start of Reo Javascript -->
     <script type="text/javascript">
