+++ portals/developer-portal/src/defaultContent/pages/api-landing/page.hbs
@@ -20,7 +20,7 @@
 <head>
   <script src='/technical-scripts/add-application-form.js' defer></script>
   <script src="/technical-scripts/oauth2-key-generation.js" defer></script>
-  <script src='/technical-scripts/platform-subscription.js' defer></script>
+  <script src='/technical-scripts/subscription.js' defer></script>
   <script src='/technical-scripts/subscription-modal.js' defer></script>
   <script src='/technical-scripts/api-agent-prompt.js' defer></script>
   <link rel="stylesheet" href="/technical-styles/agent-prompt.css">
