+++ portals/developer-portal/src/defaultContent/pages/api-landing/partials/api-detail-banner.hbs
@@ -31,7 +31,7 @@
           {{!-- Subscribe --}}
           {{#if (eq apiMetadata.apiInfo.gatewayType "wso2/api-platform")}}
             {{#if subscriptionPlans.length}}
-              {{#unless platformSubscriptions.length}}
+              {{#unless subscriptions.length}}
           <a type="button" class="common-btn-primary" href="#subscriptionPlans">Subscribe</a>
               {{/unless}}
             {{/if}}
