+++ portals/developer-portal/src/defaultContent/pages/apis/page.hbs
@@ -18,7 +18,7 @@
  -->
 <head>
   <script src='/technical-scripts/add-application-form.js' defer></script>
-  <script src='/technical-scripts/platform-subscription.js' defer></script>
+  <script src='/technical-scripts/subscription.js' defer></script>
   <script src='/technical-scripts/subscription-modal.js' defer></script>
 </head>
 <main>
