+++ portals/developer-portal/src/defaultContent/pages/subscriptions/page.hbs
@@ -19,7 +19,7 @@
 <head>
   <link rel='stylesheet' href='/technical-styles/subscriptions.css' />
   <script src='/technical-scripts/subscriptions-page.js' defer></script>
-  <script src='/technical-scripts/platform-subscription.js' defer></script>
+  <script src='/technical-scripts/subscription.js' defer></script>
 </head>
 <main>
     {{> header profile=profile baseUrl=baseUrl devportalMode=devportalMode showApiWorkflowsNav=showApiWorkflowsNav }}
