+++ portals/developer-portal/src/defaultContent/pages/subscriptions/partials/subscription-list.hbs
@@ -5,9 +5,7 @@
             <tr>
                 <th>API</th>
                 <th>Version</th>
-                <th>Type</th>
                 <th>Plan</th>
-                <th>Application</th>
                 <th>Status</th>
                 <th>Subscription Token</th>
                 <th>Actions</th>
@@ -24,32 +22,13 @@
                     <a href="{{../baseUrl}}/api/{{this.apiHandle}}">{{this.apiName}}</a>
                 </td>
                 <td>{{this.apiVersion}}</td>
-                <td>
-                    {{#if (eq this.type 'TOKEN_BASED')}}
-                    <span class="badge bg-info">Token-Based</span>
-                    {{else}}
-                    <span class="badge bg-primary">App-Based</span>
-                    {{/if}}
-                </td>
                 <td>{{this.planName}}</td>
                 <td>
-                    {{#if this.applicationName}}
-                        {{this.applicationName}}
-                    {{else}}
-                        <span class="text-muted">&mdash;</span>
-                    {{/if}}
-                </td>
-                <td>
-                    {{#if (eq this.type 'TOKEN_BASED')}}
                     <span class="badge {{#if (eq this.status 'ACTIVE')}}bg-success{{else}}bg-secondary{{/if}}">
                         {{this.status}}
                     </span>
-                    {{else}}
-                    <span class="badge bg-success">ACTIVE</span>
-                    {{/if}}
                 </td>
                 <td>
-                    {{#if (eq this.type 'TOKEN_BASED')}}
                     <div class="token-display d-flex align-items-center gap-1">
                         <code class="masked-token" id="list-token-{{this.id}}" data-revealed="false">****</code>
                         <button class="btn btn-sm btn-link p-0" onclick="toggleListTokenVisibility('{{this.id}}')" title="Reveal">
@@ -59,12 +38,8 @@
                             <i class="bi bi-clipboard"></i>
                         </button>
                     </div>
-                    {{else}}
-                    <span class="text-muted">&mdash;</span>
-                    {{/if}}
                 </td>
                 <td>
-                    {{#if (eq this.type 'TOKEN_BASED')}}
                     <div class="d-flex gap-1">
                         <button class="btn btn-sm btn-outline-warning {{#if @root.isReadOnlyMode}}disabled{{/if}}" title="{{#if (eq this.status 'ACTIVE')}}Deactivate{{else}}Activate{{/if}}"
                             {{#unless @root.isReadOnlyMode}}onclick="toggleSubscriptionStatus('{{../orgID}}', '{{this.id}}', '{{#if (eq this.status "ACTIVE")}}INACTIVE{{else}}ACTIVE{{/if}}')"{{/unless}}
@@ -72,18 +47,11 @@
                             {{#if (eq this.status 'ACTIVE')}}<i class="bi bi-pause-circle"></i>{{else}}<i class="bi bi-play-circle"></i>{{/if}}
                         </button>
                         <button class="btn btn-sm btn-outline-danger {{#if @root.isReadOnlyMode}}disabled{{/if}}" title="Delete"
-                            {{#unless @root.isReadOnlyMode}}onclick="prepareDeleteSubscription('{{../orgID}}', '{{this.id}}', 'TOKEN_BASED')"{{/unless}}
+                            {{#unless @root.isReadOnlyMode}}onclick="prepareDeleteSubscription('{{../orgID}}', '{{this.id}}')"{{/unless}}
                             {{#if @root.isReadOnlyMode}}disabled{{/if}}>
                             <i class="bi bi-trash"></i>
                         </button>
                     </div>
-                    {{else}}
-                    <button class="btn btn-sm btn-outline-danger {{#if @root.isReadOnlyMode}}disabled{{/if}}" title="Unsubscribe"
-                        {{#unless @root.isReadOnlyMode}}onclick="prepareDeleteSubscription('{{../orgID}}', '{{this.id}}', 'APP_BASED', '{{this.appId}}', '{{this.apiRefId}}')"{{/unless}}
-                        {{#if @root.isReadOnlyMode}}disabled{{/if}}>
-                        <i class="bi bi-trash"></i>
-                    </button>
-                    {{/if}}
                 </td>
             </tr>
             {{/each}}
