--- portals/developer-portal/src/openapi/handlers/platformSubscriptions.js
@@ -1,33 +1,0 @@
-/*
- * Copyright (c) 2026, WSO2 LLC. (http://www.wso2.com) All Rights Reserved.
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- */
-
-/*
- * Tag: Platform Subscriptions
- */
-const platformSubscriptionService = require('../../services/platformSubscriptionService');
-const { requireCsrfForMutatingApi } = require('../../middlewares/csrfProtection')
-const { compose } = require('./_compose');
-
-module.exports = {
-    createPlatformGatewaySubscription: compose(requireCsrfForMutatingApi, platformSubscriptionService.createPlatformGatewaySubscription),
-    listPlatformGatewaySubscriptions: platformSubscriptionService.listPlatformGatewaySubscriptions,
-    getPlatformGatewaySubscription: platformSubscriptionService.getPlatformGatewaySubscription,
-    updatePlatformGatewaySubscription: compose(requireCsrfForMutatingApi, platformSubscriptionService.updatePlatformGatewaySubscription),
-    deletePlatformGatewaySubscription: compose(requireCsrfForMutatingApi, platformSubscriptionService.deletePlatformGatewaySubscription),
-};
