+++ portals/developer-portal/src/pages/application/page.hbs
@@ -7,7 +7,7 @@
   <script src="/technical-scripts/oauth2-key-generation.js" defer></script>
   <script src="/technical-scripts/edit-application-form.js" defer></script>
   <script src="/technical-scripts/api-key-generation.js" defer></script>
-  <script src="/technical-scripts/platform-subscription.js" defer></script>
+  <script src="/technical-scripts/subscription.js" defer></script>
   <script src="/technical-scripts/subscription-modal.js" defer></script>
   <link rel="stylesheet" href="/technical-styles/subscriptions.css" />
 </head>
