+++ portals/developer-portal/src/pages/application/partials/apis.hbs
@@ -152,8 +152,8 @@
             </div>
         </div>
         <div class="row row-gap-4" style="margin-bottom: 64px;">
-            {{#if platformSubscriptions.length}}
-            <script id="token-map-data" type="application/json">{{{jsonSafePlatformSubscriptions platformSubscriptions}}}</script>
+            {{#if subscriptions.length}}
+            <script id="token-map-data" type="application/json">{{{jsonSafeSubscriptions subscriptions}}}</script>
             <script>
                 (function() {
                     var data = JSON.parse(document.getElementById('token-map-data').textContent || '[]');
@@ -182,7 +182,7 @@
                         </tr>
                     </thead>
                     <tbody>
-                        {{#each platformSubscriptions}}
+                        {{#each subscriptions}}
                         <tr class="data-row">
                             <td class="app-td" style="width: 15%;">
                                 <span onclick="window.location.href='{{@root.baseUrl}}/api/{{this.apiHandle}}'" class="text-truncate d-block" style="cursor: pointer;">{{this.apiName}}</span>
@@ -216,7 +216,7 @@
                 </table>
             </div>
             {{else}}
-            <div class="ps-3" id="no-platform-subscription">
+            <div class="ps-3" id="no-subscription">
                 <div class="info-box">
                     <div class="info-content">
                         <div class="info-title">
