+++ portals/developer-portal/src/pages/error-layout/main.hbs
@@ -32,7 +32,6 @@
     <script src='/technical-scripts/search.js' defer></script>
     <script src='/technical-scripts/filter.js' defer></script>
     <script src='/technical-scripts/common.js' defer></script>
-    <script src='/technical-scripts/subscription.js' defer></script>
 </head>
 
 <body class="main-body">
