+++ portals/developer-portal/src/pages/manage-keys/page.hbs
@@ -1,6 +1,5 @@
 <head>
   <link rel='stylesheet' href='/technical-styles/application.css' />
-  <script src='/technical-scripts/subscription.js' defer></script>
   <script src='/technical-scripts/application-dashboard.js' defer></script>
   <link rel="stylesheet" href="/technical-styles/oauth2-key-generation.css" />
   <script src="/technical-scripts/oauth2-key-generation.js" defer></script>
